up: divide logic and view
