+++ src/app.jsx
@@ -1,31 +1,8 @@
-import React, { useState } from 'react';
-import Button from './components/Button';
+import React from 'react';
+import Counter from './counter/counter';
 
-const App = () => {
-  const [accumulator, setAccumulator] = useState(0);
-  const accumulate = (amount) => {
-    const newAccumulator = accumulator + amount;
-    setAccumulator(newAccumulator);
-  };
-
-  return (
-    <div>
-      Counter
-      <div>
-        <Button onClick={() => accumulate(1)}>
-          Click me! &#40;
-          { accumulator }
-          &#41;
-        </Button>
-      </div>
-      <br />
-      <div>
-        {[1, 2, 3, 4, 5].map((number) => (
-          <Button onClick={() => accumulate(number)}>{number}</Button>
-        ))}
-      </div>
-    </div>
-  );
-};
+const App = () => (
+  <Counter />
+);
 
 export default App;
